Falcon handler: do not use deprecated `resp.body`
It is a bit tricky to keep supporting Falcon 2.0 (which is 6 year old now), but
assuming this is desirable, I changed `resp.body` to `resp.data`, since that
already exists in 2.0, and we don't need to branch on version even more.

`resp.body` has been deprecated throughout the 3.0 cycle (more than 3 years),
and it will be removed in Falcon 4.0 (a stable release is planned soon).